complete part 2: companies
Added a feature to allow filtering all companies according to name and
number of employees.

Created a middleware, convertGetAllCompaniesQueryParameters, to convert
the URL String query parameters into the correct types and values.
nameLike is decoded from URL encoding, and minEmployees and
maxEmployees are converted into integers between 0 and 2147483647.
2147483647 is the max value for PostgreSQL's integer type.

Created a helper method, sqlWhereClauseForGetCompanies, for building
the SQL Where clause when retrieving all companies.  This is appended
to the SQL statement in Company.findAll, where it is now built in
pieces.