complete part 3: change authorization
Created 2 middleware functions, ensureAdmin and ensureAdminOrSelf,
which will help make sure the currently logged-in user is an admin or
the user in the path parameter.

/companies POST, PATCH, and DELETE routes now require the currently
logged-in user to be an admin.

/users POST and GET now require user to be an admin.

/users/:username GET, PATCH, and DELETE now require the user to be an
admin or the same user as in the path parameter.

In routes/_testCommon.js, changed user "u1" property, isAdmin, to be
true.  This is done, because some routes now require the user to be an
admin.  u1 is the most used, so it is simpler to make u1 an admin than
to make and use another user.  Also created a token for user u2 for
when the currently logged-in user needs to not be an admin.