completing part 4: jobs
Adding filtering.

Allowed passing in title, minSalary, and hasEquity query parameters for
the GET route "/jobs".  title allows for case-insensitive and partial
matching.  hasEquity is a boolean and just excludes equity values of 0
if it is true.

Created a JSON schema, jobGetAllQuery.json, for validating the query
parameters, which first have to be converted from String to their
correct types.  This is done in the new middleware function,
convertGetAllJobsQueryParameters, in the file /middleware/jobs.js.

Created a helper function, sqlWhereClauseForGetJobs in /helpers/sql.js,
to factor out building the WHERE clause in the SQL statement for
retrieving all jobs.  The modular building of the WHERE clause is done
to reduce the complexity of the SQL statement.

Any integers used for the filters (minSalary) will have a maximum value
of 2147483647, which is the max value for PostgreSQL's integer data
type.