complete part 5: job applications
Added new route, POST /users/:username/jobs/:id, for allowing a user to
apply to a job.

In User.applyJob, there is no way of identifying whether the user or
job is not found, without having to perform regex on the error
details.  As a result, both are described in the error sent back to the
client.

Updated the route, GET /users/:username, to also list the job IDs of
the jobs the user has applied to.