complete further study: use enum type
Updated database schema for table applications to have a new column
called status.  This uses a new enum type for the stage a job
application is in.

In User.applyJob, changed the SQL statement to insert the "applied"
status.

Since the status isn't used elsewhere, there wasn't anything else to
change.