complete further study: add technologies for users
Added a new table, users_technologies, for attaching technologies to
users.

Created GET /:username/matchingJobs route.  This will return a list of
jobs that uses the same technologies as a specified user, with the
technologies within each job.  User info is not returned, because it
might be redundant.  There is already a GET route for that, and this
route focuses mainly on retrieving the jobs.

Created User.matchJobs to make the database call.  An extra call has to
be made to determine if a non-existent user causes no records to be
returned.  If only one call is made, the reason can be ambiguous.

In the _testCommon files, the test data is exported out so that it can
later be used for refactoring tests.