fix: add more checks for id in convertJobId
In the middleware function, convertJobId in middleware/jobs.js, the id
is only checked if it is a number.  A decimal number or a negative
would be converted without any issues, which are invalid database id
numbers.

Checks for positive integers are now added to ensure a quick fail,
instead of waiting on the database to fail the request.